+++ R/zzz.R
@@ -1,13 +1,17 @@
 #' @rawNamespace import(data.table, except = transpose)
 #' @importFrom R6 R6Class
-#' @import mlr3
+#' @importFrom mlr3 TaskClassif TaskRegr Resampling as_data_backend assert_task rsmp tsk rsmps lrn msr
 #' @import mlr3misc
 #' @import checkmate
 #' @import paradox
 #' @import ggplot2
 #' @importFrom utils globalVariables
-#' @section Additional resources:
+#' @section Main resources:
 #' * Book on mlr3: \url{https://mlr3book.mlr-org.com}
+#' * mlr3book section about spatiotemporal data: \url{https://mlr3book.mlr-org.com/spatiotemporal.html}
+#' * package vignettes: \url{https://mlr3spatiotempcv.mlr-org.com/dev/articles/}
+#'
+#' ## Miscellaneous \pkg{mlr3} content
 #' * Use cases and examples: \url{https://mlr3gallery.mlr-org.com}
 #' * More classification and regression tasks: \CRANpkg{mlr3data}
 #' * Connector to [OpenML](https://www.openml.org): \CRANpkg{mlr3oml}
@@ -22,11 +26,16 @@
 #' * Feature selection wrappers: \CRANpkg{mlr3fselect}
 #' * Interface to real (out-of-memory) data bases: \CRANpkg{mlr3db}
 #' * Performance measures as plain functions: \CRANpkg{mlr3measures}
-#' * Spatiotemporal resampling methods: \CRANpkg{mlr3spatiotempcv}
 #' * Parallelization framework: \CRANpkg{future}
 #' * Progress bars: \CRANpkg{progressr}
 #' @references
 #' `r format_bib("schratz2019")`
+#'
+#' `r format_bib("valavi2018")`
+#'
+#' `r format_bib("meyer2018")`
+#'
+#' `r format_bib("zhao2002")`
 "_PACKAGE"
 
 register_mlr3 = function() { # nocov start
@@ -51,7 +60,7 @@ register_mlr3 = function() { # nocov start
 
     # tasks --------------------------------------------------------------------
 
-    x = utils::getFromNamespace("mlr_tasks", ns = "mlr3")
+    mlr_tasks = utils::getFromNamespace("mlr_tasks", ns = "mlr3")
 
     mlr_tasks$add("ecuador", load_task_ecuador)
     mlr_tasks$add("diplodia", load_task_diplodia)
@@ -59,7 +68,7 @@ register_mlr3 = function() { # nocov start
 
     # resampling methods ---------------------------------------------------------
 
-    x = utils::getFromNamespace("mlr_resamplings", ns = "mlr3")
+    mlr_resamplings = utils::getFromNamespace("mlr_resamplings", ns = "mlr3")
     mlr_resamplings$add("spcv_block", ResamplingSpCVBlock)
     mlr_resamplings$add("spcv_buffer", ResamplingSpCVBuffer)
     mlr_resamplings$add("sptcv_cstf", ResamplingSptCVCstf)
@@ -75,8 +84,8 @@ register_mlr3 = function() { # nocov start
   }
 
   utils::globalVariables(c(
-    "row_id", "cookfarm", "ecuador", "diplodia",
-    "resampling", "task", "indicator", "fold"))
+    "row_id", "cookfarm_sample", "ecuador", "diplodia",
+    "resampling", "task", "indicator", "fold", "id", "type"))
 
 }
 
@@ -84,4 +93,13 @@ register_mlr3 = function() { # nocov start
   register_mlr3()
   setHook(packageEvent("mlr3", "onLoad"), function(...) register_mlr3(),
     action = "append")
-} # nocov end
+}
+
+.onUnload = function(libpath) { # nolint
+  event = packageEvent("mlr3", "onLoad")
+  hooks = getHook(event)
+  pkgname = vapply(hooks, function(x) environment(x)$pkgname, NA_character_)
+  setHook(event, hooks[pkgname != "mlr3spatiotempcv"], action = "replace")
+}
+
+leanify_package() # nocov end